Backend/requirements.txt
fastapi
uvicorn[standard]
torch
sentence-transformers
faiss-cpu
transformers==4.44.2
accelerate==0.33.0
numpy
scikit-learn
pydantic
